added Home function, completed Home.js Page
